Rename and relocate key files to scripts directory
Moved several core files into the 'scripts' directory for better organization. Updated all import paths to reflect the new locations of these files, ensuring consistent references across the codebase.
