+++ src/scripts/stateHandlers/default.ts
@@ -1,4 +1,4 @@
-import {StateModel} from "../../processGameFrame";
+import {StateModel} from "../processGameFrame";
 
 export default function handleProcessedGameState(processedGameState: StateModel) {
     console.log("DEFAULT")
