+++ src/scripts/workOCR.ts
